% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskanalysis.R
\name{add_location_distances}
\alias{add_location_distances}
\title{Add distances from a point location to a map.}
\usage{
add_location_distances(p, coord, distances)
}
\description{
Add distances from a point location to a map.
}
