% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animate.R
\name{animate.ca_model}
\alias{animate.ca_model}
\title{Create animated gif or video from model definitions.}
\usage{
\method{animate}{ca_model}(x, init, parms = x$parms, filename, type = "gif",
  each = 1, speed = 1, t_max = 400 * each, width = 50, height = width)
}
\arguments{
\item{x}{An object of class \code{ca_model}.}

\item{init}{Initial landscape object or cover. If cover is provided as a numerical vector of lengths \code{length(x$states)}, a new landscape object is generated.}

\item{parms}{Parameter set for simulation. If not provided, defaults are taken from model specs.}

\item{filename}{A character value specifying relative path and filename. File endings will be ignored and replaced by value in \code{type}.}

\item{type}{A character value specifying the type of output. Either "gif" or "mp4".}

\item{each}{An integer value. Distance between timesteps to be saved.}

\item{speed}{A numeric value specifying the speed of the output (defaults to 1).}

\item{t_max}{An integer value. Maximum number of timesteps. Note that one animated gif can only take up to 400 images, thus if \code{t_max > each*400} the animation will be truncated.}

\item{width}{An integer numer. Defaults to 50 and is ignored if a landscape object is provided in \code{init}.}

\item{height}{An integer numer. Defaults to 50 or to \code{width} if provided and is ignored if a landscape object is provided in \code{init}.}
}
\value{
Returns an animated gif or video.
}
\description{
Create animated gif or video from model definitions.
}
\examples{
# 1. run simulation and save full landsape at each timestep. create animated gif.

animate(life, init_landscape(c("1","0"), c(0.6,0.4), 100), filename = "life01.gif")
}

