% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animate.R
\name{animate.ca_result}
\alias{animate.ca_result}
\title{Create animated gif or video from simulation result.}
\usage{
\method{animate}{ca_result}(x, filename, type = "gif", each = 1,
  speed = 1)
}
\arguments{
\item{x}{A ca_result object (generated by \code{ca()})}

\item{filename}{A character value specifying relative path and filename. File endings will be ignored and replaced by value in \code{type}.}

\item{type}{A character value specifying the type of output. Either "gif" or "mp4".}

\item{speed}{A numeric value specifying the speed of the output (defaults to 1).}
}
\value{
Returns an animated gif or video.
}
\description{
Create animated gif or video from simulation result.
}
\details{
This requires R-package \link{[animation]} and the additional software \href{http://imagemagick.org/}{ImageMagick} (for animated .gif) and \href{https://ffmpeg.org/download.html}{FFMPEG} (for .mpeg videos).
}
\examples{
# 1. run simulation and save full landsape at each timestep. create animated gif.

l <- init_landscape(c("1","0"), c(0.6,0.4), 100)
r <- ca(l, model = life, t_max = 500, t_eval = 500, saveeach = 1)
animate(r, "life01.gif")
}
\seealso{
\link{animate.ca_model}
}

