% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca.R
\name{animate}
\alias{animate}
\title{Create animated gif.}
\usage{
animate(x, filename, type = "gif", speed = 1)
}
\arguments{
\item{x}{A ca_result object (generated by \code{ca()})}

\item{filename}{A character value specifying relative path and filename.}

\item{type}{(intended) A character value specifying the type of output. Either "gif" or "mpg".}

\item{speed}{A numeric value specifying the speed of the output (defaults to 1).}
}
\value{
Returns an animated gif file.
}
\description{
Create animated gif.
}
\details{
This requires R-package \link{[animation]} and the additional software \href{http://imagemagick.org/}{ImageMagick}
}
\examples{
# 1. run simulation and save full landsape at each timestep. create animated gif.

l <- init_landscape(c("1","0"), c(0.6,0.4), 100)
r <- ca(l, model = life, t_max = 500, t_eval = 500, saveeach = 1)
animate(r, "life01.gif")
}

