% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{as.array.ca_result}
\alias{as.array.ca_result}
\title{Transfer output of runca() into an array.}
\usage{
\method{as.array}{ca_result}(x)
}
\arguments{
\item{x}{}
}
\value{
an array with dimensions 'width', 'height' and 'snaps'.
}
\description{
Transfer output of runca() into an array.
}

