% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R, R/ca_snapsSS.R
\name{as.list.ca_result}
\alias{as.list.ca_result}
\title{Transfer output of ca() into a list of matrices.}
\usage{
\method{as.list}{ca_result}(x)

\method{as.list}{ca_result}(x)
}
\arguments{
\item{x}{}

\item{x}{}
}
\value{
a list of matrices.

a list of matrices.
}
\description{
Transfer output of ca() into a list of matrices.

Transfer output of ca() into a list of matrices.
}

