% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{as.matrix.landscape}
\alias{as.matrix.landscape}
\title{transfer landscape to matrix.}
\usage{
\method{as.matrix}{landscape}(x, as = "character")
}
\arguments{
\item{x}{A landscape object.}
}
\value{
A matrix.
}
\description{
transfer landscape to matrix.
}

