% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdeath_null.R
\docType{data}
\name{birthdeath}
\alias{birthdeath}
\title{Birth-death model}
\format{An object of class \code{ca_model} of length 6.}
\usage{
ca(l, birthdeath, parms = p)
}
\description{
Model with two states and a probability of switching between the
  two.
}
\details{
This model implements only two states (dead or alive). Each cell can 
  transition to the other depending only on a probability transition p (
  dead -> alive) or q (alive -> dead). This model serves as a null model 
  for the others provided for caspr.
}
\author{
None
}
\references{
None
}
\keyword{datasets}

