% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.R
\name{bold_stats}
\alias{bold_stats}
\title{Get statistics on marker sequences in BOLD by taxon}
\usage{
bold_stats(taxon)
}
\arguments{
\item{taxon}{taxon name}
}
\value{
a tibble
}
\description{
Get statistics on marker sequences in BOLD by taxon
}
