% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\title{Run cellular automata simulation.}
\usage{
ca(x, model = musselbed, parms = "default", t_max = 1000, t_min = 500,
  t_eval = 200, isstable = 1e-05, saveeach = 50, ...)
}
\arguments{
\item{x}{A landscape object.}

\item{model}{A valid object of class 'ca_model'. Defaults to musselbed. Valid
values are: \code{musselbed}, \code{grazing}.}

\item{parms}{A list of parameters with one or several parameters containing a
vector of parameter values. Those will be combined full-factorially using
\code{expand.grid()}. Will be checked against template parameters in
\code{model}. If not provided, template parameters will be used.}

\item{t_max}{Maximal number of timesteps. Model will be terminated even if
unstable.}

\item{t_min}{Minimal number of timesteps before stability is evaluated and
model might be terminated if transitory dynamics are surpassed.}

\item{t_eval}{Timespan that is evaluated for stability.}

\item{isstable}{Tolerance level for stability. Stability is reached if the
difference in mean cover of the primary cell state (i.e. the first in the
vector provided in \code{model$states}) over two subsequent timespans of
length \code{t_eval} is smaller than \code{isstable}.}

\item{saveeach}{Timespan between timesteps at which a full snapshot of the
landscape is saved into the output of the simulation.}

\item{...}{Parameters handed over to update function in \code{model$update}.}
}
\value{
The output is returned as a list object of class \code{ca_result}, containing a full timeseries of global and local cover as well as snapshots of the entire landscape (stored in \code{out$landscapes}).
}
\description{
Run cellular automata simulation.
}
\details{
Runs iterations of the update function \code{model$update()} on the initial landscape \code{x} until a steady state is reached (as defined by the tolerance level \code{isstable}), but max \code{t_max} timesteps. The function saves the full timeseries, i.e. a value for each timestep, of the global cover of each state as well as the average local cover of each state. Only for every \code{saveeach}th timestep, the full lattice is saved in a list within the output file (\code{result$snapshots} ).
}

