% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskanalysis.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Calculate risk from WRiMS summary and reclassified predictions.}
\usage{
calculate_risk(wrims_summ, pred_reclassified)
}
\description{
Calculate risk from WRiMS summary and reclassified predictions.
}
