% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caspr.R
\name{count.landscape}
\alias{count.landscape}
\title{Count neighbors.}
\usage{
\method{count}{landscape}(x, neighbor)
}
\arguments{
\item{x}{A landscape object.}

\item{neighbor}{A character value. The state to count.}
}
\value{
Returns a vector of the counts in the neighborhood specified by the \link{mapping}, by default the 4-cell neighborhood.
}
\description{
Count neighbors.
}

