% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskanalysis.R
\name{create_rocrate_meta}
\alias{create_rocrate_meta}
\title{Create RO-Crate metadata JSON.}
\usage{
create_rocrate_meta(title, date, caspr_version, s3_root)
}
\description{
Create RO-Crate metadata JSON.
}
