% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\name{expert_checklists}
\alias{expert_checklists}
\title{Fetch expert priority checklists from the PacMAN GitHub repository.}
\usage{
expert_checklists(summarize = TRUE, enhance = FALSE)
}
\arguments{
\item{summarize}{group by taxon and concatenate references and remarks.}

\item{enhance}{add complete taxonomy from OBIS.}
}
\value{
The expert priority checklist.
}
\description{
Fetch expert priority checklists from the PacMAN GitHub repository.
}
