% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forestgap.R
\docType{data}
\name{forestgap}
\alias{forestgap}
\title{Forest gap model}
\format{\preformatted{List of 6
 $ name  : chr "Forest Gap Model"
 $ ref   : chr "Kubo, T., Y. Iwasa and N. Furumoto. 1996. Forest spatial dynamics with gap expansion: Total gap area and gap size distribution."| __truncated__
 $ states: chr [1:2] "+" "0"
 $ cols  : chr [1:2] "#000000" "#FFFFFF"
 $ parms :List of 3
  ..$ alpha: num 0.2
  ..$ delta: num 0.17
  ..$ d    : num 0.01
 $ update:function (x, parms, subs = 10)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 21 61 1 21 1 26 61
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x6fe6088> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
forestgap
}
\description{
Forest gap model
}
\details{
We implement the version used by Kubo et al. (1996).  More details on the implementation are available at: \link{https://github.com/skefi/SpatialStress/wiki/forestgap}
}
\keyword{datasets}

