% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{fortify.ca_result}
\alias{fortify.ca_result}
\title{Fortify method for `ca_result`}
\usage{
fortify.ca_result(x)
}
\arguments{
\item{x}{a ca_result object}
}
\description{
Transform the snapshots in a ca_result object into a data.frame so it can
be plotted using ggplot.
}
\examples{
# Using the forest gap model
initial <- init_landscape(states=c("+","0"), cover=c(.5,.5))
parms <- forestgap$parms # Import defaults
parms$delta <- .15
output <- ca(initial, parms, model=forestgap, t_max=1000)

if (require(ggplot2)) {
  ggplot(fortify(output)) +
    geom_raster(aes(x,y,fill=state)) +
    facet_wrap( ~ time ) +
    scale_fill_manual(values=c('#111111','#FFFFFF'))
}
}

