% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{fortify.landscape}
\alias{fortify.landscape}
\title{#' Fortify method for `landscape`}
\usage{
fortify.landscape(x)
}
\arguments{
\item{x}{A `landscape` object}
}
\description{
Transform a `landscape` object into a data.frame so it can
be plotted using ggplot.
}
\examples{
# Using the forest gap model
initial <- init_landscape(states=c("+","0"), cover=c(.5,.5))

if (require(ggplot2)) {
  ggplot(fortify(initial)) +
    geom_raster(aes(x,y,fill=state)) +
    scale_fill_manual(values=c('#111111','#FFFFFF'))
}
}

