% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grazing.R
\docType{data}
\name{grazing}
\alias{grazing}
\title{Grazing model}
\format{\preformatted{List of 6
 $ name  : chr "Spatial Grazing Model"
 $ ref   : chr "Schneider and Kéfi 2015, in review"
 $ states: chr [1:3] "+" "0" "-"
 $ cols  : chr [1:3] "#000000" "#7F7F7F" "#FFFFFF"
 $ parms :List of 9
  ..$ del    : num 0.9
  ..$ b      : num 0.2
  ..$ c_     : num 0.2
  ..$ m0     : num 0.05
  ..$ g      : num 0.2
  ..$ r      : num 0.01
  ..$ f      : num 0.9
  ..$ d      : num 0.2
  ..$ protect: num 0.9
 $ update:function (x_old, parms_temp, subs = 10, timestep = NA)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 77 1 19 1 26 77
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x8d2da00> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
grazing
}
\description{
Grazing model
}
\details{
A model
}
\author{
Schneider and Kéfi (2015, in review)
}
\keyword{datasets}

