% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators.ca_result}
\alias{indicators.ca_result}
\title{Get spatial and temporal indicators for a ca_result object}
\usage{
\method{indicators}{ca_result}(x, spatial = TRUE, temporal = TRUE)
}
\arguments{
\item{x}{A \code{ca_result} object, output of function \code{ca()}.}

\item{spatial}{Logical. If true, function returns all spatial indicators.}

\item{temporal}{Logical. If true, function returns all temporal indicators.}
}
\value{
An object of class \code{ca_indicators} that contains a detailed
  report on all kinds of spatial and temporal indicators on the timeseries
  and snapshots.
}
\description{
Get spatial and temporal indicators for a ca_result object
}

