% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators.landscape}
\alias{indicators.landscape}
\title{Get spatial indicators for landscape object.}
\usage{
\method{indicators}{landscape}(x, state = levels(x$cells)[1], ...)
}
\arguments{
\item{x}{A landscape object.}

\item{state}{The cell state of interest. Defaults to \code{x$state[1]}.}

\item{...}{}
}
\value{
a list of spatial indicators.
}
\description{
Get spatial indicators for landscape object.
}

