% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{init_landscape}
\alias{init_landscape}
\title{Create a new landscape object, i.e. a grid for cellular automata processing.}
\usage{
init_landscape(states, cover, width = 50, height = 50)
}
\arguments{
\item{states}{A character vector containing the potential cell states.}

\item{cover}{A number vector containing the cover.}

\item{width}{An integer number.}

\item{height}{An integer number.}
}
\value{
A landscape object of dimensions \code{width} x \code{height} with
  random distribution of \code{states}, in the relative ratio given in
  \code{cover}.
}
\description{
Create a new landscape object, i.e. a grid for cellular automata processing.
}
\examples{
init_landscape(c("+","0","-"), c(0.5,0.25,0.25))
}

