% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patches.R
\name{label}
\alias{label}
\title{Labelling of unique patches.}
\usage{
label(x, state = levels(x$cells)[1])
}
\arguments{
\item{x}{A landscape object.}

\item{state}{A character vector. The state to be evaluated, defaults to the
  primary state of x (\code{state = levels(x$cells)[1]}).}
}
\value{
A landscape object containing ID numbers for each connected patch.
}
\description{
Labelling of unique patches.
}
\details{
The function is a wrapper for function \link{ConnCompLabel} of
  package 'SDMTools'.
}

