% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.R
\docType{data}
\name{life}
\alias{life}
\title{Game of Life}
\format{\preformatted{List of 7
 $ name    : chr "Random model"
 $ ref     : logi NA
 $ states  : chr [1:2] "1" "0"
 $ cols    : chr [1:2] "black" "white"
 $ interact: num [1:3, 1:3] 1 4 7 2 NA 8 3 6 9
 $ parms   :List of 2
  ..$ B: num 3
  ..$ S: num [1:2] 2 3
 $ update  :function (x_old, parms)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 28 16 39 1 16 1 28 39
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3e131a8> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
life
}
\description{
Game of Life
}
\details{
Model implementation for Conway's Game of Life.
}
\examples{
l <- init_landscape(c("1","0"), c(0.35,0.65), 150,150)

liferun <- ca(l, life)
}
\keyword{datasets}

