% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caspr.R
\name{mapping}
\alias{mapping}
\title{Create vector maps for cellular automata updating}
\usage{
mapping(width, height, boundary = "periodic", i_matrix = matrix(c(0, 1, 0,
  1, NA, 1, 0, 1, 0), ncol = 3, byrow = TRUE))
}
\arguments{
\item{width}{An integer number. The width of the lattice.}

\item{height}{An integer number. The height of the lattice.}

\item{boundary}{A character value representing the type of boundary conditions. Defaults to \code{"periodic"}.}

\item{i_matrix}{A matrix defining the interaction matrix to be evaluated. Defaults to 4-cell neighborhood.}
}
\value{
Creates a map of the landscape object to translate it into a vector with boundaries (\code{x_with_border}) and another one to back-translate it to a vector without boundaries  (\code{x_to_evaluate}) into the global environment. This will automatically be called within \code{ca()}.
}
\description{
Create vector maps for cellular automata updating
}

