% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caspr.R
\name{neighbors}
\alias{neighbors}
\title{Count neighbors.}
\usage{
neighbors(x, state = levels(x$cells)[1])
}
\arguments{
\item{x}{A landscape object.}

\item{state}{A character value. The state to count. Defaults to the primary
  cell state, i.e. the first state given in \code{levels(x$cells)}.}
}
\value{
Returns a vector of the counts in the neighborhood of each cell in
  the landscape. Neighborhood is specified by the global
  \code{\link{mapping}} objects, by default the 4-cell neighborhood.
}
\description{
Count neighbors.
}
\details{
This function is vectorised by using the mapping objects defined by
  a previous call of \code{\link{mapping}}.
}
\seealso{
For counting the neighborhood of \emph{one} target file, e.g. when
  doing sequential updating of randomly drawn cells, the function
  \code{\link{neighborsij}} is slightly faster. However, for such updating
  processes we recommend using C++ via \code{Rcpp} (see implementation of
  the predator prey model, \code{predprey.R}, for an example).
}

