% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caspr.R
\name{neighborsij}
\alias{neighborsij}
\title{Count neighbors of one cell.}
\usage{
neighborsij(x, state = levels(x$cells)[1], i = NULL, j = NULL,
  select = 1:prod(x$dim))
}
\description{
Count neighbors of one cell.
}

