% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\name{obis_checklist}
\alias{obis_checklist}
\title{Fetch a checklist from OBIS.}
\usage{
obis_checklist(wrims = TRUE, area = "south_pacific")
}
\arguments{
\item{wrims}{limit checklist to WRiMS taxa.}

\item{area}{limit checklist to area, options are south_pacific and fiji.}
}
\value{
An OBIS checklist.
}
\description{
Fetch a checklist from OBIS.
}
