% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms_timeseries.R
\name{parms_timeseries}
\alias{parms_timeseries}
\title{Create sequence of parameters over time.}
\usage{
parms_timeseries(parms, t_max)
}
\arguments{
\item{parms}{A list of parameters. Each list entry can be either a single value or a vector representing the sequence of parameter values over time (in years). This is inflated by the requested time in t_max.}

\item{t_max}{An integer number. The number of years to evaluate the timeseries.}
}
\description{
Create sequence of parameters over time.
}
\examples{

p <- list(  
  r = 1.0,  # max. regeneration rate of plants
  b = seq(0.4,0.1, length = 5),  # environmental quality
  sigma_b = 0.1, # random annual variation of environmental quality
  f = 0.8,  # local facilitation
  alpha = 0, # water runoff
  K = 0.9, # carrying capacity of the system
  c = 0.2, # local competition
  m = 0.05, # intrinsic mortality of plants (inverse of av. lifespan)
  v = 0.8, # attractant-decoy
  p = 0.99, # associational resistance
  L = rep(c(0,10), each = 1), # Livestock density
  q = 0, # hill exponent of functional response
  h = 50, # handling time 
  a = 0.2 # attack rate of livestock
) 


plist <- parms_timeseries(p, 5)

l <- init_landscape(c("1","0"), cover = c(0.5,0.5), width = 100)

ca(l, livestock, plist, t_max = 5, saveeach = 5)

}

