% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patches.R
\name{patches}
\alias{patches}
\title{Count patch sizes}
\usage{
patches(x, state = levels(x$cells)[1])
}
\arguments{
\item{x}{A landscape object.}

\item{state}{A character value. The state of interest to be counted.}
}
\value{
Returns a vector of patch sizes (number of cells) occuring in the landscape.
}
\description{
Count patch sizes
}
\details{
Might be replaced by a faster version implemented in C++.
}

