% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialwarningsS3.R
\name{patchsizes}
\alias{patchsizes}
\title{Wrapper functions for package 'spatialwarnings'}
\usage{
indicators(l)
indicators_fitpsd(l)
patchsizes(l)
}
\description{
A set of wrapper functions that apply the spatial analysis tools provided by 'spatialwarnings' to caspr landscape objects.
}

