% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca.R
\name{plot.ca_result}
\alias{plot.ca_result}
\title{plot method for 'ca_result' objects}
\usage{
\method{plot}{ca_result}(x, plotstates = c(TRUE, FALSE, FALSE),
  snapshots = FALSE, cols = x$model$cols, lwd = 1, ...)
}
\arguments{
\item{x}{An output object obtained from running function \code{ca()}.}

\item{plotstates}{A logical vector of the same length as
\code{x$model$states}. TRUE activates plotting of other cell state populations.}

\item{cols}{A vector of valid colors to replace the default color vector.}

\item{lwd}{Line width for timeseries plot.}

\item{...}{Parameters handed to standard plot function, e.g. `bty`, `xlab`.
See \link{par}.}
}
\description{
plot method for 'ca_result' objects
}
\details{
This prompts a (series of) summary plot for the simulation run. By
default this is a timeseries of the primary cell state (i.e. the first state
given in \code{x$model$states}).
}

