% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{plot.landscape}
\alias{plot.landscape}
\title{Plotting an objects of class "landscape"}
\usage{
\method{plot}{landscape}(x, cols = "auto", grid = FALSE, axis = FALSE,
  add = FALSE, ani = FALSE, ...)
}
\arguments{
\item{x}{A landscape object.}

\item{cols}{A color vector. If \code{"auto"}, then a grayscale vector will be
used.}

\item{grid}{If TRUE, plot a grid over the cells. Defaults to FALSE.}

\item{axis}{If TRUE, plot x and y axis with coordinates.  Defaults to FALSE.}

\item{add}{If TRUE, no primary plot will be called. The change in cells will
be plotted over an existing plot. Used for animated plotting in the screen
plotting device.}

\item{ani}{If TRUE, an adjustment constant is added when producing a pixel
accurate png or gif file. Required when the function is used to plot
animated figures.}
}
\value{
A landscape object of dimensions \code{width} x \code{height} with
  random distribution of \code{states}, in the relative ratio given in
  \code{cover}.
}
\description{
Plotting an objects of class "landscape"
}
\examples{
onj <- init_landscape(c("+","0","-"), c(0.5,0.25,0.25))
plot(obj)
}

