% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskanalysis.R
\name{plot_location_distances}
\alias{plot_location_distances}
\title{Plot distances from a point location on a map.}
\usage{
plot_location_distances(coord, distances)
}
\description{
Plot distances from a point location on a map.
}
