% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_seas.R
\name{regional_seas}
\alias{regional_seas}
\title{Variation on cartomisc::regional_seas with a small buffer applied before joining.}
\usage{
regional_seas(
  x,
  group,
  dist = units::set_units(30, km),
  density = units::set_units(0.1, 1/km)
)
}
\description{
Variation on cartomisc::regional_seas with a small buffer applied before joining.
}
