% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/steady.R
\name{steady}
\alias{steady}
\title{Test for steady state dynamics of timeseries.}
\usage{
steady(i, result, steadyparms = list(t_eval = 200, accept = 0.001),
  returnvalue = FALSE)
}
\arguments{
\item{i}{}

\item{result}{}

\item{steadyparms}{}

\item{returnvalue}{}
}
\value{
a binary value. TRUE, if steady state is reached. FALSE, if still in
  transient dynamics.
}
\description{
Test for steady state dynamics of timeseries.
}

