% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{summary.landscape}
\alias{summary.landscape}
\title{Summary of landscape object.}
\usage{
\method{summary}{landscape}(x)
}
\arguments{
\item{x}{A landscape object}
}
\value{
A list containing the object dimensions, the number of cells in the different cell states, the global cover of the different cell states, the average local cover of the different cell states.
}
\description{
Summary of landscape object.
}
\examples{
obj <- init_landscape(c("+","0","-"), c(0.5,0.25,0.25))
summary(obj)
}

