% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_template.R
\docType{data}
\name{template}
\alias{template}
\title{Template for model objects}
\format{An object of class \code{ca_model} of length 6.}
\usage{
ca(l, template, parms = p)
}
\description{
One sentence description. Potential cell states.
}
\details{
Wordy description about the model details, transition probabilities, biological assumptions.
}
\author{
Author of original model
}
\references{
Original reference and further references
}
\keyword{datasets}

