% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\docType{data}
\name{wkt_south_pacific}
\alias{wkt_south_pacific}
\title{WKT string for the South Pacific}
\format{
An object of class \code{character} of length 1.
}
\usage{
wkt_south_pacific
}
\description{
WKT string for the South Pacific
}
\keyword{datasets}
