% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrims.R
\name{wrims_checklist}
\alias{wrims_checklist}
\title{Get the complete WRiMS checklist. This may not be complete due to missing macroalgae.}
\usage{
wrims_checklist()
}
\description{
Get the complete WRiMS checklist. This may not be complete due to missing macroalgae.
}
