% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrims.R
\name{wrims_distribution}
\alias{wrims_distribution}
\title{Get WRiMS distribution records.}
\usage{
wrims_distribution(aphiaid, simplify = FALSE, cleanup = FALSE)
}
\description{
Get WRiMS distribution records.
}
