% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrims.R
\name{wrims_summary_for_aphiaid}
\alias{wrims_summary_for_aphiaid}
\title{Create summary of WRiMS distribution.}
\usage{
wrims_summary_for_aphiaid(
  aphiaid,
  unknown = 0.75,
  add_background = FALSE,
  simplify = 0.1,
  hole_max_area = 1e+08,
  union_buffer = 0.01,
  buffer = 1
)
}
\description{
Create summary of WRiMS distribution.
}
