% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{as.landscape.matrix}
\alias{as.landscape.matrix}
\title{transfer matrix to landscape.}
\usage{
\method{as.landscape}{matrix}(x)
}
\arguments{
\item{x}{A matrix object with factorial content.}
}
\value{
A landscape object.
}
\description{
transfer matrix to landscape.
}

