% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/birthdeath_null.R
\docType{data}
\name{birthdeath}
\alias{birthdeath}
\title{Birth-death model}
\format{\preformatted{List of 6
 $ name  : chr "Random birth-death model"
 $ ref   : logi NA
 $ states: chr [1:2] "+" "0"
 $ cols  : chr [1:2] "black" "white"
 $ parms :List of 2
  ..$ death: num 0.5
  ..$ birth: num 0.5
 $ update:function (x_old, parms)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 39 22 52 1 22 1 39 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3c75220> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
ca(l, birthdeath, parms = p)
}
\description{
Model with two states and a probability of switching between the
  two.
}
\details{
This model implements only two states (dead or alive). Each cell can
  transition to the other depending only on a probability transition p (
  dead -> alive) or q (alive -> dead). This model serves as a null model
  for the others provided for caspr.
}
\author{
None
}
\references{
None
}
\keyword{datasets}

