% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count.landscape}
\alias{count.landscape}
\title{Count neighbors.}
\usage{
\method{count}{landscape}(x, neighbor)
}
\arguments{
\item{x}{A landscape object.}

\item{neighbor}{A character value. The state to count.}
}
\value{
Returns a vector of the counts in the neighborhood specified by the \code{\link{mapping}}, by default the 4-cell neighborhood.
}
\description{
Count neighbors.
}

