% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grazing.R
\docType{data}
\name{grazing}
\alias{grazing}
\title{Grazing model}
\format{\preformatted{List of 6
 $ name  : chr "Spatial Grazing Model"
 $ ref   : chr "Schneider and Kefi 2015, in review"
 $ states: chr [1:3] "+" "0" "-"
 $ cols  : chr [1:3] "#000000" "#7F7F7F" "#FFFFFF"
 $ parms :List of 9
  ..$ del: num 0.9
  ..$ b  : num 0.5
  ..$ c_ : num 0.2
  ..$ m0 : num 0.05
  ..$ g  : num 0.2
  ..$ r  : num 0.01
  ..$ f  : num 0.9
  ..$ d  : num 0.1
  ..$ p  : num 1
 $ update:function (x_old, parms_temp, subs = 10, timestep = NA)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 85 19 136 1 19 1 85 136
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x400ff18> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
ca(l, grazing)
}
\arguments{
\item{del}{local seed dispersal}

\item{b}{environmental quality}

\item{c_}{global competition}

\item{m0}{intrinsic mortality}

\item{g}{grazing pressure}

\item{r}{regeneration rate of degraded cells}

\item{f}{local facilitation}

\item{d}{intrinsic degradation rate}

\item{p}{associational resistance against grazing}
}
\description{
A spatially-explicit model of grazing pressure in drylands.
}
\details{
The model builds upon a published model by Kefi et al. 2007. Spatial
  models of vegetation cover so far have considered grazing mortality a
  rather constant pressure, affecting all plants equally, regardless of their
  position in space. In the known models it usually adds as a constant to the
  individual plant risk (Kefi et al 2007 TPB). However, grazing has a strong
  spatial component: Many plants in rangelands invest in protective
  structures such as thorns or spines, or develop growth forms that reduce
  their vulnerability to grazing. Therefore, plants growing next to each
  other benefit from the protection of their neighbors.

  Such \strong{associational resistance} is widely acknowledged in vegetation
  ecology but hardly integrated in models as a cause for spatially
  heterogenous grazing pressure. It also renders the plant mortality density
  dependent, which has important impacts on the bistability of the system.

  The model investigates how the assumption of spatially heterogeneous
  pressure alters the bistability properties and the response of spatial
  indicators of catastrophic shifts.

  The model knows three different cell states: occupied by vegetation
  \code{"+"}, empty but fertile \code{"0"} and degraded \code{"-"}.
  Transitions between cell states are only possible between vegetated and
  empty (by the processes of plant 'death' and 'recolonization') and between
  empty and degraded (by 'degradation' and 'regeneration').

  To account for the spatially heterogeneous impacts of grazing due to
  associational resistance, we assumed that a plant's vulnerability to
  grazers decreases with the proportion of occupied neighbors,  $q_{+|+}$.
  The individual probability of dying is therefore defined as

  \deqn{	w_{ \left\{ +,0 \right\} }  = m_0 + g_0 \left( 1 - q_{+|+} \right)}

  where the additional mortality due to grazing is maximized to \eqn{g_0} if
  a plant has no vegetated neighbor (i.e., \eqn{q_{+|+} = 0}) and gradually
  reduces to 0 with an increasing fraction of occupied neighbors,
  \eqn{q_{+|+}}.
}
\examples{
l <- init_landscape(c("+","0","-"), c(0.6,0.2,0.2), width = 50) # create initial landscape
p <- list(del = 0.9, b = 0.4, c_ = 0.2, m0 = 0.05, g = 0.25, r = 0.01, f = 0.9, d = 0.1, p = 1)   # set parameters
r <- ca(l, model = grazing, parms = p, t_max = 100)    # run simulation
}
\author{
Florian D. Schneider and Sonia Kefi (2015, in revision)
}
\seealso{
\href{https://github.com/cascade-wp6/2015_schneider_kefi}{project on GitHub}

Other models: \code{\link{forestgap}}; \code{\link{life}};
  \code{\link{livestock}}; \code{\link{musselbed}};
  \code{\link{predprey}}
}
\keyword{datasets}

