% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/livestock.R
\docType{data}
\name{livestock}
\alias{livestock}
\title{Livestock resilience model}
\format{\preformatted{List of 6
 $ name  : chr "Livestock resilience model"
 $ ref   : logi NA
 $ states: chr [1:2] "1" "0"
 $ cols  : chr [1:2] "black" "white"
 $ parms :List of 13
  ..$ r    : num 1
  ..$ b    : num 0.2
  ..$ f    : num 0.9
  ..$ alpha: num 0
  ..$ K    : num 0.8
  ..$ c    : num 0.1
  ..$ m    : num 0.05
  ..$ v    : num 0
  ..$ p    : num 0.8
  ..$ L    : num 6
  ..$ q    : num 0
  ..$ h    : num 50
  ..$ a    : num 0.2
 $ update:function (x_old, parms, subs = 12)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 21 92 1 21 1 50 92
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4101528> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
ca(l, livestock, parms = p)
}
\arguments{
\item{b}{environmental quality}

\item{m}{intrinsic plant mortality, i.e. inverse of av. lifespan}

\item{r}{max. regeneration rate of plants}

\item{K}{carrying capacity of the landscape (i.e. global competition)}

\item{f}{local facilitation}

\item{c}{local competition}

\item{L}{livestock density}

\item{a}{attack rate of livestock}

\item{h}{handling time of livestock}

\item{q}{hill-exponent of livestock}

\item{p}{associational resistance against livestock grazing}

\item{v}{attractant-decoy effect of plants to livestock}
}
\description{
A spatially-explicit model of resilience in arid rangelands.
}
\details{
An unpublished, generalized model of positive and negative local feedbacks in arid rangelands, including mechanisms such as local facilitation, competition, associational resistance and attractant-decoy.
}
\author{
F.D. Schneider
}
\keyword{datasets}

