% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/musselbed.R
\docType{data}
\name{musselbed}
\alias{musselbed}
\title{Mussel bed model}
\format{\preformatted{List of 6
 $ name  : chr "Mussel Disturbance Model"
 $ ref   : chr "Guichard et al 2003, American Naturalist, Vol. 161, pp. 889–904"
 $ states: chr [1:3] "+" "0" "-"
 $ cols  : chr [1:3] "#000000" "#7F7F7F" "#FFFFFF"
 $ parms :List of 3
  ..$ r    : num 0.4
  ..$ d    : num 0.9
  ..$ delta: num 0.01
 $ update:function (x_old, parms_temp, delta = 0.2, subs = 10, timestep = NA)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 19 21 66 1 21 1 19 66
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x8cc0a08> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
musselbed
}
\description{
Mussel bed model
}
\author{
Guichard et al 2003
}
\keyword{datasets}

