% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predprey.R
\docType{data}
\name{predprey}
\alias{predprey}
\title{Predator-prey model}
\format{\preformatted{List of 6
 $ name  : chr "Predator-prey Gap Model"
 $ ref   : chr "Pascual, M., Roy, M., Guichard, F., & Flierl, G. (2002). Cluster size distributions: signatures of self-organization in spatial"| __truncated__
 $ states: chr [1:3] "f" "s" "0"
 $ cols  : chr [1:3] "gray50" "black" "white"
 $ parms :List of 3
  ..$ betaf: num 0.333
  ..$ betas: num 0.1
  ..$ delta: num 0.333
 $ update:function (x, parms, subs = prod(x$dim))  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 39 20 58 1 20 1 39 58
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2cf40c8> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
predprey
}
\arguments{
\item{betaf}{}

\item{betas}{}

\item{delta}{}
}
\description{
Predator-prey model
}
\author{
Pascual, M., Roy, M., Guichard, F., & Flierl, G. (2002). Cluster size
  distributions: signatures of self-organization in spatial ecologies.
  Philosophical Transactions of the Royal Society of London. Series B,
  Biological Sciences, 357(1421), 657–666.
  http://doi.org/10.1098/rstb.2001.0983
}
\keyword{datasets}

