% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_template.R
\docType{data}
\name{template}
\alias{template}
\title{Template for model objects}
\format{\preformatted{List of 6
 $ name  : chr "Random model"
 $ ref   : logi NA
 $ states: chr [1:2] "1" "0"
 $ cols  : chr [1:2] "black" "grey50"
 $ parms :List of 2
  ..$ death: num 0.5
  ..$ birth: num 0.5
 $ update:function (x_old, parms)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 27 20 38 1 20 1 27 38
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x404f748> 
 - attr(*, "class")= chr "ca_model"
}}
\usage{
ca(l, template, parms = p)
}
\description{
One sentence description. Potential cell states.
}
\details{
Wordy description about the model details, transition probabilities, biological assumptions.
}
\author{
Author of original model
}
\references{
Original reference and further references
}
\keyword{datasets}

